WIP - Delete Popover for assessments
- refactor file-delete to allow for assessments as well.
- update logic on user-dash.ts

+++ src/pages/user-dashboard/user-dashboard.ts
@@ -1,8 +1,9 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, EventEmitter } from '@angular/core';
 import { IonicPage, NavController, NavParams, PopoverController } from 'ionic-angular';
 import { AuthService } from "../../services/auth.service";
 import { AssessmentService } from "../../services/assessment.service";
 import { TopbarComponent } from "../../components/topbar/topbar";
+import { FileDeleteComponent } from '../../components/file-delete/file-delete';
 import { SettingsPage } from "../settings/settings";
 import { QuestionsPage } from "../questions/questions";
 import { DashboardPage } from "../dashboard/dashboard";
@@ -97,8 +98,6 @@ export class UserDashboardPage {
 		}
   }
 
-
-
 	async getSharedAssessments() {
 		var user;
 		if (this.auth.currentUser()) {
@@ -177,9 +176,27 @@ export class UserDashboardPage {
 	toggleMine = () => {this.showMine = !this.showMine;}
 	toggleShared = () => {this.showShared = !this.showShared;}
 
+  /** 
+  *   launch delete popover, pass assessment type
+  *   create an emitter to recieve user response from popover,
+  *   if emitter returns truthy, go use assessment service delete,
+  *   & remove from view
+  *   assessmentId: the assessmentId of the assessment to be deleted
+  */
 	async handleDeleting(assessmentId){
-		var observe =  await this.assessmentService.deleteAssessment(assessmentId);
-		observe.subscribe((data => this.removeAssessmentFromPage(assessmentId)) );
+    var emitter =  new EventEmitter<any>();
+    emitter.subscribe(deleteFile => { 
+      if (deleteFile) {
+        this.assessmentService.deleteAssessment(assessmentId)
+        .then(a => a.toPromise())
+        .catch( err => console.error('cant resolve to Promise'))
+        .then(p => this.removeAssessmentFromPage(assessmentId))
+        .catch( err => console.error('cant remove from page'));
+      }
+    });
+
+    this.popOver.create(FileDeleteComponent, {emitter: emitter, typeToDelete: 'assessment'})
+                .present();
 	}
 
 	presentAddTeamMembersPopOver(assessmentId){
